feat: create routes and use in component
