+++ src/api/profileAPI/changeProfileInformation.ts
@@ -1,18 +1,26 @@
 import toast from 'react-hot-toast';
-import { Address } from '../types/user/formData';
+import { ProfileFormData } from '../../schemas/profile.schema';
 
-export type Action =
-  | { action: 'changeAddress'; address: Address; addressId: string }
-  | { action: 'setDefaultShippingAddress'; addressId?: string }
-  | { action: 'setDefaultBillingAddress'; addressId?: string };
-
-interface changeAddressProps {
-  userToken?: string;
+interface changeProfileInforamtionProps {
+  customer: ProfileFormData;
+  userToken: string;
   version: number;
-  actions: Action[];
 }
 
-export const changeAddress = async ({ actions, userToken, version }: changeAddressProps) => {
+export const changeProfileInformation = async ({
+  customer,
+  userToken,
+  version,
+}: changeProfileInforamtionProps) => {
+  const { firstName, lastName, dateOfBirth, email } = customer;
+
+  const actions = [
+    { action: 'setFirstName', firstName },
+    { action: 'setLastName', lastName },
+    { action: 'setDateOfBirth', dateOfBirth },
+    { action: 'changeEmail', email },
+  ];
+
   try {
     const response = await fetch(`https://api.europe-west1.gcp.commercetools.com/dino-land/me`, {
       method: 'POST',
@@ -21,8 +29,8 @@ export const changeAddress = async ({ actions, userToken, version }: changeAddre
         'Content-Type': 'application/json',
       },
       body: JSON.stringify({
-        actions,
         version,
+        actions,
       }),
     });
 
@@ -32,7 +40,7 @@ export const changeAddress = async ({ actions, userToken, version }: changeAddre
         toast.error(errorData.message);
         throw new Error(errorData.message);
       }
-      throw new Error('Error change address');
+      throw new Error('Error change profile information');
     } else {
       toast.success('Change successfull!');
     }
