+++ src/routes/routes.ts
@@ -1,0 +1,9 @@
+export const ROUTES = {
+  HOME: '/',
+  ABOUT: '/about',
+  LOGIN: '/login',
+  PROFILE: '/profile',
+  REGISTRATION: '/registration',
+  CATALOG: '/catalog',
+  CART: '/cart',
+};
